# perf(TestComparer): Optimize GC memory allocation.