--- a/src/Uno.UI.TestComparer/Uno.UI.TestComparer.csproj
+++ b/src/Uno.UI.TestComparer/Uno.UI.TestComparer.csproj
@@ -3,12 +3,13 @@
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
 		<TargetFramework>net472</TargetFramework>
-	  <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-  </PropertyGroup>
-
+		<TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+		<AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
+		<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+	</PropertyGroup>
+	
 	<ItemGroup>
-	  <None Remove="Properties\launchSettings.json" />
+		<None Remove="Properties\launchSettings.json" />
 	</ItemGroup>	
 
 	<ItemGroup>
@@ -20,18 +21,12 @@
 		<PackageReference Include="NUnit.Engine" Version="3.11.1" />
 		<PackageReference Include="Refit" Version="6.0.94" />
 		<PackageReference Include="Mono.Options" Version="6.6.0.161" />
+		<PackageReference Include="SkiaSharp" Version="2.88.0-preview.140" />
 		<PackageReference Include="System.IO.Compression" Version="4.3.0" />
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Folder Include="Properties\" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <Reference Include="PresentationCore" />
-	  <Reference Include="PresentationFramework" />
-	  <Reference Include="System.Windows.Forms" />
-	  <Reference Include="WindowsBase" />
+		<Folder Include="Properties\" />
 	</ItemGroup>
 
 </Project>