@jeromelaban 2021-10-05
Thanks for the contribution! While the change is indeed allocating a bit less, it's not having a significant impact on CI machines (most of the time is spent on disk and network I/O, not on CPU time).

Did you notice a change when running locally ?