@workgroupengineering 2021-10-05
You're right, I was just focused on memory usage, not much can be done on network i / o.

After this change fc45648858c295467579402eb13ab40802aa896b, however, I have optimized the use of the cpu by about 30%.

It sure is always better than nothing.

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19042.1237 (20H2/October2020Update)
Intel Core i7-7700K CPU 4.20GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
  [Host]   : .NET Framework 4.8 (4.8.4400.0), X64 RyuJIT
  ShortRun : .NET Framework 4.8 (4.8.4400.0), X64 RyuJIT

Job=ShortRun  InvocationCount=1  IterationCount=3  
LaunchCount=1  UnrollFactor=1  WarmupCount=3  

```
|   Method |                pairs | write |     Mean |    Error |  StdDev |      Min |      Max |   Median | Ratio | Rank |
|--------- |--------------------- |------ |---------:|---------:|--------:|---------:|---------:|---------:|------:|-----:|
| **Uno(WPF)** | **Syste(...)Pair] [53]** | **False** |  **92.50 s** | **14.942 s** | **0.819 s** |  **91.56 s** |  **93.08 s** |  **92.85 s** |  **1.00** |    **2** |
|     Skia | Syste(...)Pair] [53] | False |  83.90 s |  1.619 s | 0.089 s |  83.82 s |  83.99 s |  83.88 s |  0.91 |    1 |
|          |                      |       |          |          |         |          |          |          |       |      |
| **Uno(WPF)** | **Syste(...)Pair] [53]** |  **True** | **133.62 s** | **36.284 s** | **1.989 s** | **131.87 s** | **135.78 s** | **133.20 s** |  **1.00** |    **2** |
|     Skia | Syste(...)Pair] [53] |  True |  87.81 s |  9.523 s | 0.522 s |  87.21 s |  88.19 s |  88.02 s |  0.66 |    1 |